controller is returning user details instead of user (username/password)

+++ CoderDojo/src/main/java/com/skilldistillery/coderdojo/repositories/UserDetailRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.skilldistillery.coderdojo.entities.UserDetail;
 
-public interface UserDetailRepository extends JpaRepository<UserDetail, Integer> {
+public interface UserDetailRepository extends JpaRepository<UserDetail, Long> {
 	UserDetail findByUserId(long id);
 }
